% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tikzsave.R
\name{compile_tikzpicture}
\alias{compile_tikzpicture}
\title{Wraps and compiles a tikzpicture}
\usage{
compile_tikzpicture(
  filename,
  packages = NULL,
  pre_begin_document = NULL,
  recompile = TRUE,
  create_png = FALSE
)
}
\arguments{
\item{filename}{Character. Filename of a tikzpicture in \code{.tex} file}

\item{packages}{Vector of strings. If no \verb{\\} is detected, then this is
assumed to be an absolute file paths to a \code{.sty} file to
include in preamble. Default is my \code{paper.sty} and \code{math.sty} in
\code{inst/tikzsave}.}

\item{pre_begin_document}{Verbatim text to include right before
\verb{\\begin\{document\}}}

\item{recompile}{Logical. If FALSE, does not update if pdf date
is newer than the tex date.}

\item{create_png}{Logical. If true, will also output a png in addition to
the pdf.}
}
\value{
Invisibly returns \code{TRUE} if compilation succeeds.
}
\description{
Takes the output of \code{tikzDevice::tikz} with \code{standalone = FALSE}, wraps into
a latex document and then compiles using \code{latexmk}, cleaning up as it goes.
}
