% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful.R
\name{copy_vec}
\alias{copy_vec}
\title{Prints vector with each item on a new line}
\usage{
copy_vec(vec)
}
\arguments{
\item{vec}{Vector you want to print}
}
\description{
Prints vector with each item on a new line
}
