% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.R
\name{create_sparse}
\alias{create_sparse}
\title{Create sparse matrix from triplet of coordinates and value}
\usage{
create_sparse(i, j, x = 1)
}
\arguments{
\item{i}{A vector of row membership}

\item{j}{A vector of column membership}

\item{x}{A vector of the entries that should go in (i, j)}
}
\value{
A sparse matrix of class \code{dgCMatrix} from the Matrix package
}
\description{
Create sparse matrix from triplet of coordinates and value
}
\examples{
i <- sample(letters, 10, replace = TRUE)
j <- sample(1:100, 10, replace = TRUE)
create_sparse(i, j)

}
