% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_table_body.R
\name{extract_body_etable}
\alias{extract_body_etable}
\title{Extract table body from \code{\link[fixest:etable]{fixest::etable}}}
\usage{
extract_body_etable(tab)
}
\arguments{
\item{tab}{The result of \code{\link[fixest:etable]{fixest::etable}} with
\code{tex = TRUE}}
}
\description{
Extract table body from \code{\link[fixest:etable]{fixest::etable}}
}
\section{Examples}{
\if{html}{\out{<div class="r">}}\preformatted{est <- fixest::feols(hp ~ 1 + mpg, data = mtcars)
tab <- fixest::etable(est, tex = TRUE, fitstat = c("n", "ar2"))
}\if{html}{\out{</div>}}\preformatted{#> \\begin\{tabular\}\{lc\}
#> \\tabularnewline\\midrule\\midrule
#> Dependent Variable:&hp\\\\
#> Model:&(1)\\\\
#> \\midrule \\emph\{Variables\}&  \\\\
#> (Intercept) & 324.1$^\{***\}$\\\\
#>   &(27.43)\\\\
#> mpg & -8.830$^\{***\}$\\\\
#>   &(1.310)\\\\
#> \\midrule \\emph\{Fit statistics\}&  \\\\
#> Observations & 32\\\\
#> Adjusted R$^2$ & 0.58919\\\\
#> \\midrule\\midrule\\multicolumn\{2\}\{l\}\{\\emph\{IID standard-errors in parentheses\}\}\\\\
#> \\multicolumn\{2\}\{l\}\{\\emph\{Signif. Codes: ***: 0.01, **: 0.05, *: 0.1\}\}\\\\
#> \\end\{tabular\}
}\if{html}{\out{<div class="r">}}\preformatted{# export to file with file option of cat
cat(extract_body_etable(tab))
}\if{html}{\out{</div>}}\preformatted{#> Dependent Variable:&hp\\\\
#> 
#> Model:&(1)\\\\
#> 
#> \\midrule \\emph\{Variables\}&  \\\\
#> 
#> (Intercept) & 324.1$^\{***\}$\\\\
#>   &(27.43)\\\\
#> mpg & -8.830$^\{***\}$\\\\
#>   &(1.310)\\\\
#> 
#> \\midrule \\emph\{Fit statistics\}&  \\\\
#> 
#> Observations & 32\\\\
#> Adjusted R$^2$ & 0.58919\\\\
}\if{html}{\out{<div class="r">}}\preformatted{est_fe <- fixest::feols(hp ~ 1 + mpg | cyl, data = mtcars)
tab_fe <- fixest::etable(est_fe, tex = TRUE, fitstat = c("n", "ar2"))
}\if{html}{\out{</div>}}\preformatted{#> \\begin\{tabular\}\{lc\}
#> \\tabularnewline\\midrule\\midrule
#> Dependent Variable:&hp\\\\
#> Model:&(1)\\\\
#> \\midrule \\emph\{Variables\}&  \\\\
#> mpg & -3.327$^\{*\}$\\\\
#>   &(1.114)\\\\
#> \\midrule \\emph\{Fixed-effects\}&  \\\\
#> cyl & Yes\\\\
#> \\midrule \\emph\{Fit statistics\}&  \\\\
#> Observations & 32\\\\
#> Adjusted R$^2$ & 0.70855\\\\
#> \\midrule\\midrule\\multicolumn\{2\}\{l\}\{\\emph\{Clustered (cyl) standard-errors in parentheses\}\}\\\\
#> \\multicolumn\{2\}\{l\}\{\\emph\{Signif. Codes: ***: 0.01, **: 0.05, *: 0.1\}\}\\\\
#> \\end\{tabular\}
}\if{html}{\out{<div class="r">}}\preformatted{cat(extract_body_etable(tab_fe))
}\if{html}{\out{</div>}}\preformatted{#> Dependent Variable:&hp\\\\
#> 
#> Model:&(1)\\\\
#> 
#> \\midrule \\emph\{Variables\}&  \\\\
#> 
#> mpg & -3.327$^\{*\}$\\\\
#>   &(1.114)\\\\
#> 
#> \\midrule \\emph\{Fixed-effects\}&  \\\\
#> 
#> cyl & Yes\\\\
#> 
#> \\midrule \\emph\{Fit statistics\}&  \\\\
#> 
#> Observations & 32\\\\
#> Adjusted R$^2$ & 0.70855\\\\
}
}

