% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_table_body.R
\name{extract_body_gt}
\alias{extract_body_gt}
\title{Extract table body from \code{\link[gt::gt]{gt::gt}}}
\usage{
extract_body_gt(tab)
}
\arguments{
\item{tab}{The result of \code{\link[gt::gt]{gt::gt}} or after manually
calling \code{as.character(gt::as_latex(tab))}}
}
\description{
Extract table body from \code{\link[gt::gt]{gt::gt}}
}
\section{Examples}{

\subsection{gt}{\if{html}{\out{<div class="r">}}\preformatted{tab <- gt::gt(head(mtcars))

# export to file with file option of cat
cat(extract_body_gt(tab))
}\if{html}{\out{</div>}}\preformatted{#> NA
}
}

\subsection{gtsummary}{\if{html}{\out{<div class="r">}}\preformatted{library(gtsummary)
tab <- tbl_summary(
  mtcars[,c("cyl", "hp", "mpg")],
  statistic = list(all_continuous() ~ "\{mean\} (\{sd\})"),
  by = "cyl"
) \%>\%
  # remove footnotes
  modify_footnote(update = everything() ~ NA)

cat(extract_body_gt(tab))
}\if{html}{\out{</div>}}\preformatted{#> NA
}
}
}

