% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fact_to_sparse.R
\name{fact_to_sparse}
\alias{fact_to_sparse}
\title{Creates sparse 0/1 matrix for factor variables}
\usage{
fact_to_sparse(df)
}
\arguments{
\item{df}{dataframe containing factor columns that will be turned into sparse matrix.
Note that the variables aren't required to be factor variables themselves.
Make sure not to include other variables in df}
}
\value{
\code{\link[Matrix::sparseMatrix]{Matrix::sparseMatrix}} with 0/1 matrix
}
\description{
This function is a really fast and memory efficient way to create a
0/1 matrix from a factor variable. This is helpful for clustered standard errors.

Original from: https://stackoverflow.com/a/23042363/8785011
}
\examples{
n <- 1e6
df <- data.frame(
 x = factor(sample(c("A", "B", "C"), n, TRUE)),
	y = factor(sample(c("D", "E"), n, TRUE))
)
mat <- fact_to_sparse(df)

}
