% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_utils.R
\name{fill_column}
\alias{fill_column}
\title{Conditionally fill column with custom function}
\usage{
fill_column(gtobj, columns, FUN, ...)
}
\arguments{
\item{gtobj}{Object from \code{gt::gt()}}

\item{columns}{The names of the columns that are to be targeted.}

\item{FUN}{Function that takes options \code{...} and returns a vector of color values.
Note that the first two arguments should be \code{data} which will be \code{gtobj[["_data"]]}
and \code{columns} which will be passed from the argument}
}
\description{
Conditionally fill column with custom function
}
\section{Examples}{
\if{html}{\out{<div class="r">}}\preformatted{color_if <- function(data, columns, var, value) \{
  color <- rep("transparent", nrow(data))
  color[data[[var]] == value] <- "red"
  color
\}

gt(mtcars) \%>\%
  fill_column(column = "mpg", FUN = color_if, var = "cyl", val = 6)
}\if{html}{\out{</div>}}
}

