% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_theme.R
\name{gt_theme_kyle}
\alias{gt_theme_kyle}
\title{Custom gt theme which borrows heavily from Thomas Mock}
\usage{
gt_theme_kyle(data, ...)
}
\arguments{
\item{data}{\code{gt} object. Pipe into this function to add theme}

\item{...}{Additional arguments to pass to \code{gt::tab_options}}
}
\description{
Custom gt theme which borrows heavily from Thomas Mock
}
\examples{

library(gt)

mtcars[1:5, ] |>
  gt() |>
  gt_theme_kyle() |>
  tab_header(title = make_gt_title("mtcars Dataset")) |>
  tab_source_note(source_note = "Data from mtcars in R")

}
