% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documenting.R
\name{h1}
\alias{h1}
\title{Section headers}
\usage{
h1(txt)
}
\arguments{
\item{txt}{Text you want to wrap in my header format}
}
\description{
Section headers
}
\details{
\code{h1} has the following output that is automatically copied to the clipboard

\verb{# ------------------------------------------------------------------------------}

\verb{# txt}

\verb{# ------------------------------------------------------------------------------}
}
