% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{haversine}
\alias{haversine}
\title{Compute Haversine distance between two points}
\usage{
haversine(lat1, lon1, lat2, lon2)
}
\arguments{
\item{lat1}{Latitude for starting point}

\item{lon1}{Longitude for starting point}

\item{lat2}{Latitude for ending point}

\item{lon2}{Longitude for ending point}
}
\value{
Double of distance between coordinate pairs in km
}
\description{
Compute Haversine distance between two points
}
