% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documenting.R
\name{header2}
\alias{header2}
\title{Subsection headers}
\usage{
header2(txt)
}
\arguments{
\item{txt}{Text you want to wrap in my header format}
}
\description{
Subsection headers
}
\details{
\code{h2} has the following output that is automatically copied to the clipboard

\verb{# ---- txt ---------------------------------------------------------------------}
}
