% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documenting.R
\name{introblock}
\alias{introblock}
\title{Intro block for top of script}
\usage{
introblock(filename, author = "Kyle Butts, CU Boulder Economics", txt = "")
}
\arguments{
\item{filename}{Name of file (including .R)}

\item{author}{Name of author. Defaults to me}

\item{txt}{Optional, text description of file}
}
\description{
Intro block for top of script
}
\details{
\code{introblock} has the following output that is automatically copied to the clipboard

\verb{## filename.R ------------------------------------------------------------------}
\verb{## Kyle Butts, CU Boulder Economics}
\verb{## }
\verb{## txt}
}
