% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_utils.R
\name{kyle_textbox}
\alias{kyle_textbox}
\title{Nice looking annotations using `ggtext``}
\usage{
kyle_textbox(
  label,
  x,
  y,
  size = 10,
  width = unit(0.2, "npc"),
  text.color = tailwind_color("zinc-800"),
  fill = "white",
  box.color = tailwind_color("zinc-300"),
  box.size = unit(0.5, "pt"),
  box.r = unit(4, "pt"),
  box.padding = unit(c(6, 6, 6, 6), "pt"),
  hjust = 0,
  vjust = 1
)
}
\arguments{
\item{label}{Text to be displayed}

\item{x}{x coordinate of the label}

\item{y}{y coordinate of the label}

\item{size}{Text size in points}

\item{width}{width of textbox}

\item{text.color}{color of text}

\item{fill}{background color}

\item{hjust}{horizontal justification}

\item{vjust}{vertical justification}

\item{border.color}{color of border}

\item{linewidth}{width of border}

\item{label.r}{radius of rounded corners}

\item{label.padding}{padding around the label}
}
\value{
A ggplot layer object with a textbox annotation
}
\description{
Nice looking annotations using `ggtext``
}
