% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_tools.R
\name{preview_html}
\alias{preview_html}
\title{Preview html string in RStudio viewer pane}
\usage{
preview_html(html)
}
\arguments{
\item{html}{A valid html string. Can be a \code{character}, \code{html} object from
\code{gt::html}, or \code{shiny.tag} from \code{htmltools}.}
}
\description{
This function lets you preview html in the Viewer tab of RStudio. This
is helpful when creating things with custom css using htmltools
(or modifying gt tables)
}
\section{Examples}{
\if{html}{\out{<div class="r">}}\preformatted{library(fontawesome)
library(htmltools)
more_or_less_css <- css(
    font.size = "13px",
    width = "80px",
    font.family = "arial",
    display = "flex",
    flex.direction = "column",
    justify.content = "center",
    text.align = "center",
    padding.right = "5px",
    color = "#999"
)

more_or_less <- div(
    style = more_or_less_css,
    div(
        style = css(padding.bottom = "20px"),
        fontawesome::fa("angle-double-up"),
        p("Increasing")
    ),
    div(
        style = css(padding.top = "40px"),
        p("Decreasing"),
        fontawesome::fa("angle-double-down")
    )
)

# preview_html(more_or_less)
}\if{html}{\out{</div>}}
}

