% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xaringan_utils.R
\name{print_rmd}
\alias{print_rmd}
\title{Print Xaringan .Rmd/.html file using \code{pagedown::chrome_print}}
\usage{
print_rmd(file, pdf_file = NULL)
}
\arguments{
\item{file}{The full file path to .Rmd/.html file}

\item{pdf_file}{Specify custom pdf output location. Otherwise the same
location of file will be used}
}
\description{
Print Xaringan .Rmd/.html file using \code{pagedown::chrome_print}
}
