% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_utils.R
\name{pt_to_mm}
\alias{pt_to_mm}
\title{Convert a numeric vector specified in pt to mm}
\usage{
pt_to_mm(x)
}
\arguments{
\item{x}{Numeric vector of units specified in pt}
}
\value{
Numeric vector of units specified in mm
}
\description{
Convert a numeric vector specified in pt to mm
}
