% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_parallel_distm_C}
\alias{rcpp_parallel_distm_C}
\title{Parallel calculates distance between two matrices of lat/long.}
\usage{
rcpp_parallel_distm_C(x, y)
}
\arguments{
\item{x}{matrix of coordinates. In the same format as st_coordinates output (X, Y)}

\item{y}{matrix of coordinates. In the same format as st_coordinates output (X, Y)}
}
\value{
returns matrix of distances in miles
}
\description{
Parallel calculates distance between two matrices of lat/long.
}
