% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_parallel_distm_C_min}
\alias{rcpp_parallel_distm_C_min}
\title{Find closest unit from group y}
\usage{
rcpp_parallel_distm_C_min(x, y)
}
\arguments{
\item{x}{matrix of coordinates. In the same format as st_coordinates output (X, Y)}

\item{y}{matrix of coordinates. In the same format as st_coordinates output (X, Y)}
}
\value{
Vector. Contains index of closest unit in y
}
\description{
Find closest unit from group y
}
