% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_parallel_distm_C_min_nonself}
\alias{rcpp_parallel_distm_C_min_nonself}
\title{Find nearest neighbor of x (non-self)}
\usage{
rcpp_parallel_distm_C_min_nonself(x, y)
}
\arguments{
\item{x}{matrix of coordinates. In the same format as st_coordinates output (X, Y)}

\item{y}{Same as x.}
}
\value{
Vector. Contains index of closest unit in y
}
\description{
Find nearest neighbor of x (non-self)
}
