% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_parallel_nearest_facility}
\alias{rcpp_parallel_nearest_facility}
\title{Find closest unit from group y and measure distance}
\usage{
rcpp_parallel_nearest_facility(x, y)
}
\arguments{
\item{x}{Matrix of coordinates. In the same format as st_coordinates output (X, Y)}

\item{y}{Matrix of coordinates. In the same format as st_coordinates output (X, Y)}
}
\value{
Matrix. 2 columns. First column is index of closest unit in y.
Second column is distance
}
\description{
Find closest unit from group y and measure distance
}
