% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_colors.R
\name{scale_fill_kyle}
\alias{scale_fill_kyle}
\title{Fill scale for kyle colors}
\usage{
scale_fill_kyle(discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{discrete}{Boolean to indicate if color aesthetic is discrete.}

\item{reverse}{Boolean to indicate whether palette should be reversed.}

\item{...}{Additional arguments passed to \code{discrete_scale} or
\code{scale_fill_viridis_c}, depending on the value of \code{discrete}.}
}
\description{
Fill scale for kyle colors
}
