% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_functions.R
\name{st_dist_rcpp}
\alias{st_dist_rcpp}
\title{RcppParallel version of \code{st_distance}}
\usage{
st_dist_rcpp(x, y = x, unit = "mi")
}
\arguments{
\item{x}{Object of class sf/sfc or a matrix of coordinates. If a
matrix of coordinates, use column order from \code{st_coordinates()}.}

\item{y}{Optional. Object of class sf/sfc or a matrix of coordinates. If a
matrix of coordinates, use column order from \code{st_coordinates()}.
If not included, x is used.}

\item{unit}{Either "mi", "km", or "m". Sets the output distance.}
}
\value{
Matrix of distances of dimension n_x by n_y.
}
\description{
RcppParallel version of \code{st_distance}
}
\examples{
library(sf)
nc <- st_read(system.file("shape/nc.shp", package="sf"))
st_dist_rcpp(nc[1:3,])

nc_long <- nc[rep(seq_len(nrow(nc)), 5), ]
sf::st_distance(nc_long)
st_dist_rcpp(nc_long)

}
