% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_functions.R
\name{st_geometry_to_lat_long}
\alias{st_geometry_to_lat_long}
\title{Convert geometry column to \code{longitude}/\code{latitude} columns}
\usage{
st_geometry_to_lat_long(df, geom_column = "geometry")
}
\arguments{
\item{df}{data.frame with a \code{geometry} column}

\item{geom_column}{String containing the name of the geometry column.
Default is geometry}
}
\description{
Convert geometry column to \code{longitude}/\code{latitude} columns
}
