% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_functions.R
\name{st_multiply_sgbp}
\alias{st_multiply_sgbp}
\title{Multiplies a sparse geometry binary predicate object (e.g. from \code{sf::st_within}) and a vector}
\usage{
st_multiply_sgbp(sgbp, mat)
}
\arguments{
\item{sgbp}{\code{sgbp} object from an \code{sf} function, e.g. \code{sf::st_within}}

\item{mat}{Vector/Matrix you want to perform sgbp \%*\% mat.}
}
\description{
Performs sgbp \%\emph{\% mat, where sgbp is a 0/1 matrix.
The main use case is when you want to sum a variable from points into a
geometry. For example, sgbp \%}\% rep(1, n) will give you the number within
each geometry.
}
