% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\docType{data}
\name{tailwind}
\alias{tailwind}
\title{Tailwind color palette}
\format{
An object of class \code{character} of length 11.
}
\usage{
tailwind
}
\description{
Source: \url{https://tailwindcss.com/docs/customizing-colors}
}
\keyword{datasets}
