% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailwind_colors.R
\name{tailwind_color}
\alias{tailwind_color}
\title{Pass string of TailwindCSS color and get hex'}
\usage{
tailwind_color(color)
}
\arguments{
\item{color}{A color named}
}
\value{
A string containing the hex code for the corresponding color
}
\description{
Colors from https://tailwindcss.com/docs/customizing-colors
}
