% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tikzsave.R
\name{tikzsave}
\alias{tikzsave}
\title{Save a tikzpicture and then compile to pdf}
\usage{
tikzsave(
  filename,
  plot = ggplot2::last_plot(),
  packages = NULL,
  pre_begin_document = NULL,
  recompile = TRUE,
  create_png = FALSE,
  quiet = TRUE,
  ...
)
}
\arguments{
\item{filename}{File name to create on disk.}

\item{plot}{Plot to save, defaults to last plot displayed.}

\item{packages}{Vector of strings. If no \verb{\\} is detected, then this is
assumed to be an absolute file paths to a \code{.sty} file to
include in preamble. Default is my \code{paper.sty} and \code{math.sty} in
\code{inst/tikzsave}.}

\item{pre_begin_document}{Verbatim text to include right before
\verb{\\begin\{document\}}}

\item{recompile}{Logical. Since \code{tikzDevice::tikz} caches effectively,
sometimes the \code{.pdf} file will be newer than the saved \code{.tex} file.
If \code{FALSE}, this will skip recomipilation in these cases.
Default is \code{TRUE}.}

\item{create_png}{Logical. If true, will also output a png in addition to
the pdf.}

\item{quiet}{Logical. If \code{TRUE} (default), do not display info messages.}

\item{...}{Passed to \code{ggplot2::ggsave}}
}
\value{
Invisibly returns filename
}
\description{
Uses \code{ggplot2::ggsave} with \code{tikzDevice::tikz} to create the tikz figure
and then compiles using \code{latexmk}.
}
