% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tikzsave.R
\name{compile_tikzpicture}
\alias{compile_tikzpicture}
\title{Wraps and compiles a tikzpicture}
\usage{
compile_tikzpicture(filename, packages = NULL, recompile = TRUE)
}
\arguments{
\item{filename}{Character. Filename of a tikzpicture in \code{.tex} file}

\item{packages}{Vector of strings. Absolute file paths to \code{.sty} files to
include in preamble. Default is my \code{paper.sty} and \code{math.sty} in
\code{inst/tikzsave}.}

\item{recompile}{Logical. If FALSE, does not update if pdf date
is newer than the tex date.}
}
\value{
invisible
}
\description{
Takes the output of \code{tikzDevice::tikz} with \code{standalone = FALSE}, wraps into
a latex document and then compiles using \code{latexmk}, cleaning up as it goes.
}
