% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dist_spherical_cosine}
\alias{dist_spherical_cosine}
\title{Spherical Distance}
\usage{
dist_spherical_cosine(lon_x, lat_x, lon_y, lat_y)
}
\arguments{
\item{lon_x}{Longitude for starting point}

\item{lat_x}{Latitude for ending point}

\item{lon_y}{Longitude for ending point}
}
\value{
Double of distance between points in m
}
\description{
Spherical Distance
}
