% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_table_body.R
\name{extract_body_etable}
\alias{extract_body_etable}
\title{Extract table body from \code{fixest::etable}}
\usage{
extract_body_etable(tab)
}
\arguments{
\item{tab}{The result of \code{fixest::etable} with
\code{tex = TRUE}}
}
\description{
Extract table body from \code{fixest::etable}
}
\section{Examples}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{est <- fixest::feols(hp ~ 1 + mpg, data = mtcars)
tab <- fixest::etable(est, tex = TRUE, fitstat = c("n", "ar2"))

# export to file with file option of cat
cat(extract_body_etable(tab))
#> NA

est_fe <- fixest::feols(hp ~ 1 + mpg | cyl, data = mtcars)
tab_fe <- fixest::etable(est_fe, tex = TRUE, fitstat = c("n", "ar2"))
cat(extract_body_etable(tab_fe))
#> NA
}\if{html}{\out{</div>}}
}

