% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_utils.R
\name{fill_column}
\alias{fill_column}
\title{Conditionally fill column with custom function}
\usage{
fill_column(gtobj, columns, FUN, ...)
}
\arguments{
\item{gtobj}{Object from \code{gt::gt()}}

\item{columns}{The names of the columns that are to be targeted.}

\item{FUN}{Function that returns a vector of color values.
Note that the first two arguments should be \code{data} which will be \code{gtobj[["_data"]]}
and \code{columns} which passes \code{columns}. Additional optinos can be passed from \code{...}}
}
\description{
Conditionally fill column with custom function
}
\details{
If you have columns you want to reference in FUN that you don't want
in the final table, use \code{gt::cols_hide()}
}
\section{Examples}{


Simple example that colors if a variable in the data equals some value

\if{html}{\out{<div class="sourceCode r">}}\preformatted{color_if <- function(data, columns, var, value) \{
  color <- rep("transparent", nrow(data))
  color[data[[var]] == value] <- "red"
  color
\}

gt(mtcars) |>
  fill_column(column = "mpg", FUN = color_if, var = "cyl", val = 6)
}\if{html}{\out{</div>}}
}

