% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.R
\name{get_min_year}
\alias{get_min_year}
\title{From an indicator for treatment, retrieve start year.}
\usage{
get_min_year(y, t)
}
\arguments{
\item{y}{Vector of years}

\item{t}{Vector of 0/1 for treatment}
}
\value{
Returns earliest y for which t == 1
}
\description{
This function returns the earliest year where an indicator == 1.
If indicator never equals zero, return Inf.
This is most useful with \code{group_by(unit) |> mutate(g = ...)} or
\code{DT(j = g := ..., by = unit)} to convert between D and g variable in panel data.
}
