% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_tools.R
\name{html2R}
\alias{html2R}
\title{Convert raw html to htmltools/shiny tags}
\usage{
html2R(raw_html)
}
\arguments{
\item{raw_html}{Character vector of html text}
}
\value{
Character string of htmltools:: version of html
}
\description{
Note that this copies text to your clipboard
}
\examples{
raw1 <- htmltools::div(
  class = "max-w-6xl",
  htmltools::div(
    class = "prose", style = "width:100\%;",
    htmltools::h1("Header"),
    "prose"
  ),
  htmltools::div(
    class = "prose",
    htmltools::h2("Smaller Header"),
    htmltools::p(
      "Testing ",
      htmltools::span(class = "italic", "this function")
    ),
    htmltools::img(src = "cat.png", "custom-attr" = 1)
  )
) |>
  as.character()

html2R(raw1)

raw2 <- r"(
<div class="container-fluid max-w-6xl">
  <div class="prose" style="width:100\%;">
    <h1>Header</h1>
    prose
  </div>
  <div class="prose">
    <h2>Smaller Header</h2>
    <p>Testing <span class="italic">this function</span></p>
    <img src="cat.png" custom-attr="1"/>
  </div>
</div>)"

html2R(raw2) |> cat()

}
