% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{kyle_color}
\alias{kyle_color}
\title{Get color from my palette by name}
\usage{
kyle_color(color = NULL)
}
\arguments{
\item{color}{Character(s) of colors}
}
\value{
Vector of colors. If color is null, returns the full palette.
Otherwise, the vector is the corresponding colors
}
\description{
Get color from my palette by name
}
