% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_colors.R
\docType{data}
\name{kyle_colors}
\alias{kyle_colors}
\title{Set of colors for graphics that I like}
\format{
An object of class \code{character} of length 7.
}
\usage{
kyle_colors
}
\description{
They are accessible: https://projects.susielu.com/viz-palette?colors=\link[="#002c55","#b3114b","#5c4cbf","#158ea6","#fb7185","#77c669","#ffc517"]{\%22#002c55\%22,\%22#b3114b\%22,\%22#5c4cbf\%22,\%22#158ea6\%22,\%22#fb7185\%22,\%22#77c669\%22,\%22#ffc517\%22}&backgroundColor=\%22white\%22&fontColor=\%22black\%22&mode=\%22achromatopsia\%22
}
\details{
The order of the colors go from dark to light
}
\keyword{datasets}
