% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_helpers.R
\name{png_to_grob}
\alias{png_to_grob}
\title{Create image grob from png and output the width and height}
\usage{
png_to_grob(source)
}
\arguments{
\item{source}{name of image to read}
}
\description{
This takes a png file and converts it to a grob. This is useful in
combination with
\code{\link[ggplot2::annotation_custom()]{ggplot2::annotation_custom()}}.o
}
\examples{

library(ggplot2)

img <- png_to_grob(system.file("img", "Rlogo.png", package="png"))

ggplot(mtcars, aes(x = mpg, y = wt, color = factor(cyl))) +
  geom_point() +
  annotation_custom(img, xmin = 25, xmax = 30, ymin = 4, ymax =  5)
}
