% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_parallel_nearest_facility_nonself}
\alias{rcpp_parallel_nearest_facility_nonself}
\title{Find nearest neighbor of x and measure distance (non-self)}
\usage{
rcpp_parallel_nearest_facility_nonself(x, y)
}
\description{
@param x. matrix of coordinates. In the same format as st_coordinates output (X, Y)
@param y. Same as x.
}
\details{
@return Matrix. 2 columns. First column is index of closest unit in y.
Second column is distance

@export
}
