% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_functions.R
\name{st_dist_rcpp}
\alias{st_dist_rcpp}
\title{RcppParallel version of \code{st_distance}}
\usage{
st_dist_rcpp(x, y = x, unit = "mi")
}
\arguments{
\item{x.}{Object of class sf/sfc or a matrix of coordinates. If a
matrix of coordinates, use column order from \code{st_coordinates()}.}

\item{y.}{Optional. Object of class sf/sfc or a matrix of coordinates. If a
matrix of coordinates, use column order from \code{st_coordinates()}.
If not included, x is used.}

\item{unit.}{Either "mi", "km", or "m". Sets the output distance.}
}
\value{
Matrix of distances of dimension n_x by n_y.
}
\description{
RcppParallel version of \code{st_distance}
}
\section{Examples}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(sf)
nc <- st_read(system.file("shape/nc.shp", package="sf"))
#> Reading layer `nc' from data source 
#>   `/Library/Frameworks/R.framework/Versions/4.2-arm64/Resources/library/sf/shape/nc.shp' 
#>   using driver `ESRI Shapefile'
#> Simple feature collection with 100 features and 14 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -84.32385 ymin: 33.88199 xmax: -75.45698 ymax: 36.58965
#> Geodetic CRS:  NAD27
st_dist_rcpp(nc[1:3,])
#> Warning in st_point_on_surface.sfc(x): st_point_on_surface may not give correct
#> results for longitude/latitude data
#> > Distance in miles. To use kilometers use option `unit == 'km'`
#>              [,1]         [,2]         [,3]
#> [1,] 2.383768e-13 2.050102e+01 4.471554e+01
#> [2,] 2.050102e+01 9.055578e-14 2.517342e+01
#> [3,] 4.471554e+01 2.517342e+01 1.907309e-13
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{nc_long <- nc[rep(seq_len(nrow(nc)), 5), ]
# massive speedup
microbenchmark::microbenchmark(
 sf = sf::st_distance(nc_long),
  parallel = st_dist_rcpp(nc_long),
  times = 1
)
#> Warning in microbenchmark::microbenchmark(sf = sf::st_distance(nc_long), : less
#> accurate nanosecond times to avoid potential integer overflows
#> Warning in st_point_on_surface.sfc(x): st_point_on_surface may not give correct
#> results for longitude/latitude data
#> > Distance in miles. To use kilometers use option `unit == 'km'`
#> Unit: milliseconds
#>      expr          min           lq         mean       median           uq
#>        sf 77455.156970 77455.156970 77455.156970 77455.156970 77455.156970
#>  parallel     9.174037     9.174037     9.174037     9.174037     9.174037
#>           max neval
#>  77455.156970     1
#>      9.174037     1
}\if{html}{\out{</div>}}
}

