% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_functions.R
\name{st_extract_within}
\alias{st_extract_within}
\title{Extract information from geometry that a point x falls in}
\usage{
st_extract_within(x, y, var)
}
\arguments{
\item{x}{\code{sf} object consisting of points}

\item{y}{\code{sf} object consisting of geometries}

\item{var}{character of variable name you wish to extract from \code{y}}
}
\description{
For each point in x, identify which geometry in y the point falls within
and extract value of variable \code{var}. This is good, for example, if you have
points and you want to extract information about the county/census tract/etc.
}
\section{Examples}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(sf)
nc <- st_read(system.file("shape/nc.shp", package="sf"))
#> Reading layer `nc' from data source 
#>   `/Library/Frameworks/R.framework/Versions/4.2-arm64/Resources/library/sf/shape/nc.shp' 
#>   using driver `ESRI Shapefile'
#> Simple feature collection with 100 features and 14 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -84.32385 ymin: 33.88199 xmax: -75.45698 ymax: 36.58965
#> Geodetic CRS:  NAD27
x <- st_as_sf(data.frame(id = 1:10, geometry = st_sample(nc, 10)))
x$CNTY_ID = st_extract_within(x, nc, "CNTY_ID")
x
#> Simple feature collection with 10 features and 2 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -83.06399 ymin: 34.98672 xmax: -77.19209 ymax: 36.38153
#> Geodetic CRS:  NAD27
#>    id                   geometry CNTY_ID
#> 1   1 POINT (-77.19209 36.27941)    1833
#> 2   2 POINT (-81.90775 35.53488)    1964
#> 3   3 POINT (-81.46535 36.01603)    1932
#> 4   4 POINT (-79.17714 36.01998)    1907
#> 5   5 POINT (-77.78655 34.98672)    2100
#> 6   6 POINT (-77.71775 35.91597)    1928
#> 7   7 POINT (-78.62955 36.38153)    1840
#> 8   8 POINT (-83.06399 35.22193)    2039
#> 9   9 POINT (-79.99384 35.42833)    2044
#> 10 10 POINT (-78.85485 35.14009)    2090
}\if{html}{\out{</div>}}
}

