% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_functions.R
\name{st_extract_within}
\alias{st_extract_within}
\title{Extract information from geometry that a point x falls in}
\usage{
st_extract_within(x, y, var)
}
\arguments{
\item{x}{\code{sf} object consisting of points}

\item{y}{\code{sf} object consisting of geometries}

\item{var}{character of variable name you wish to extract from \code{y}}
}
\description{
For each point in x, identify which geometry in y the point falls within
and extract value of variable \code{var}. This is good, for example, if you have
points and you want to extract information about the county/census tract/etc.
}
\section{Examples}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(sf)
nc <- st_read(system.file("shape/nc.shp", package="sf"))
#> Reading layer `nc' from data source 
#>   `/Library/Frameworks/R.framework/Versions/4.3-arm64/Resources/library/sf/shape/nc.shp' 
#>   using driver `ESRI Shapefile'
#> Simple feature collection with 100 features and 14 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -84.32385 ymin: 33.88199 xmax: -75.45698 ymax: 36.58965
#> Geodetic CRS:  NAD27
x <- st_as_sf(data.frame(id = 1:10, geometry = st_sample(nc, 10)))
x$CNTY_ID = st_extract_within(x, nc, "CNTY_ID")
x
#> Simple feature collection with 10 features and 2 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -81.17788 ymin: 34.77648 xmax: -76.12275 ymax: 36.47084
#> Geodetic CRS:  NAD27
#>    id                   geometry CNTY_ID
#> 1   1 POINT (-76.67583 34.77648)    2156
#> 2   2 POINT (-80.98233 35.95033)    1947
#> 3   3 POINT (-79.05387 35.44651)    2026
#> 4   4 POINT (-79.70627 36.47084)    1839
#> 5   5 POINT (-77.27123 34.94471)    2091
#> 6   6 POINT (-76.12275 35.86011)    1963
#> 7   7 POINT (-78.03558 34.81841)    2100
#> 8   8 POINT (-81.17788 35.88598)    1950
#> 9   9 POINT (-79.71549 35.17716)    2044
#> 10 10 POINT (-76.81477 35.34936)    2004
}\if{html}{\out{</div>}}
}

