% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_functions.R
\name{st_extract_within}
\alias{st_extract_within}
\title{Extract information from geometry that a point x falls in}
\usage{
st_extract_within(x, y, var)
}
\arguments{
\item{x}{\code{sf} object consisting of points}

\item{y}{\code{sf} object consisting of geometries}

\item{var}{character of variable name you wish to extract from \code{y}}
}
\description{
For each point in x, identify which geometry in y the point falls within
and extract value of variable \code{var}. This is good, for example, if you have
points and you want to extract information about the county/census tract/etc.
}
\section{Examples}{
\if{html}{\out{<div class="r">}}\preformatted{library(sf)
}\if{html}{\out{</div>}}\preformatted{#> Linking to GEOS 3.8.1, GDAL 3.2.1, PROJ 7.2.1
}\if{html}{\out{<div class="r">}}\preformatted{nc <- st_read(system.file("shape/nc.shp", package="sf"))
}\if{html}{\out{</div>}}\preformatted{#> Reading layer `nc' from data source `/Library/Frameworks/R.framework/Versions/4.1/Resources/library/sf/shape/nc.shp' using driver `ESRI Shapefile'
#> Simple feature collection with 100 features and 14 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -84.32385 ymin: 33.88199 xmax: -75.45698 ymax: 36.58965
#> Geodetic CRS:  NAD27
}\if{html}{\out{<div class="r">}}\preformatted{x <- st_as_sf(data.frame(id = 1:10, geometry = st_sample(nc, 10)))
x$CNTY_ID = st_extract_within(x, nc, "CNTY_ID")
x
}\if{html}{\out{</div>}}\preformatted{#> Simple feature collection with 10 features and 2 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -81.45147 ymin: 33.97097 xmax: -76.22612 ymax: 36.51061
#> Geodetic CRS:  NAD27
#>    id                   geometry CNTY_ID
#> 1   1 POINT (-79.34913 35.65129)    1973
#> 2   2 POINT (-78.87864 34.51243)    2150
#> 3   3 POINT (-78.65762 33.97097)    2232
#> 4   4  POINT (-79.0486 36.15368)    1907
#> 5   5 POINT (-76.22612 36.21375)    1848
#> 6   6 POINT (-77.76955 34.46412)    2185
#> 7   7  POINT (-78.5587 36.51061)    1840
#> 8   8 POINT (-80.74044 35.14085)    2041
#> 9   9 POINT (-77.04036 36.24152)    1833
#> 10 10 POINT (-81.45147 35.98629)    1932
}
}

