% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_functions.R
\name{st_extract_within}
\alias{st_extract_within}
\title{Extract information from geometry that a point x falls in}
\usage{
st_extract_within(x, y, var)
}
\arguments{
\item{x}{\code{sf} object consisting of points}

\item{y}{\code{sf} object consisting of geometries}

\item{var}{character of variable name you wish to extract from \code{y}}
}
\description{
For each point in x, identify which geometry in y the point falls within
and extract value of variable \code{var}. This is good, for example, if you have
points and you want to extract information about the county/census tract/etc.
}
\section{Examples}{
\if{html}{\out{<div class="r">}}\preformatted{library(sf)
nc <- st_read(system.file("shape/nc.shp", package="sf"))
}\if{html}{\out{</div>}}\preformatted{#> Reading layer `nc' from data source `/Library/Frameworks/R.framework/Versions/4.1/Resources/library/sf/shape/nc.shp' using driver `ESRI Shapefile'
#> Simple feature collection with 100 features and 14 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -84.32385 ymin: 33.88199 xmax: -75.45698 ymax: 36.58965
#> Geodetic CRS:  NAD27
}\if{html}{\out{<div class="r">}}\preformatted{x <- st_as_sf(data.frame(id = 1:10, geometry = st_sample(nc, 10)))
x$CNTY_ID = st_extract_within(x, nc, "CNTY_ID")
x
}\if{html}{\out{</div>}}\preformatted{#> Simple feature collection with 10 features and 2 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -84.29255 ymin: 34.88861 xmax: -76.93521 ymax: 36.48406
#> Geodetic CRS:  NAD27
#>    id                   geometry CNTY_ID
#> 1   1 POINT (-76.93521 35.56519)    2004
#> 2   2 POINT (-84.29255 35.08652)    2088
#> 3   3 POINT (-77.12606 35.24603)    2119
#> 4   4 POINT (-80.42304 34.88861)    2095
#> 5   5 POINT (-79.05568 35.24326)    2030
#> 6   6 POINT (-79.67945 35.98406)    1903
#> 7   7 POINT (-78.40544 35.52534)    1989
#> 8   8 POINT (-77.57178 36.48406)    1832
#> 9   9 POINT (-83.96951 35.34739)    2056
#> 10 10  POINT (-78.8531 35.08022)    2090
}
}

