% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_functions.R
\name{st_extract_within}
\alias{st_extract_within}
\title{Extract information from geometry that a point x falls in}
\usage{
st_extract_within(x, y, var)
}
\arguments{
\item{x}{\code{sf} object consisting of points}

\item{y}{\code{sf} object consisting of geometries}

\item{var}{character of variable name you wish to extract from \code{y}}
}
\description{
For each point in x, identify which geometry in y the point falls within
and extract value of variable \code{var}. This is good, for example, if you have
points and you want to extract information about the county/census tract/etc.
}
\section{Examples}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(sf)
nc <- st_read(system.file("shape/nc.shp", package="sf"))
#> Reading layer `nc' from data source 
#>   `/Library/Frameworks/R.framework/Versions/4.2-arm64/Resources/library/sf/shape/nc.shp' 
#>   using driver `ESRI Shapefile'
#> Simple feature collection with 100 features and 14 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -84.32385 ymin: 33.88199 xmax: -75.45698 ymax: 36.58965
#> Geodetic CRS:  NAD27
x <- st_as_sf(data.frame(id = 1:10, geometry = st_sample(nc, 10)))
x$CNTY_ID = st_extract_within(x, nc, "CNTY_ID")
x
#> Simple feature collection with 10 features and 2 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -81.14601 ymin: 33.97557 xmax: -76.51494 ymax: 36.31645
#> Geodetic CRS:  NAD27
#>    id                   geometry CNTY_ID
#> 1   1    POINT (-77.625 36.2204)    1846
#> 2   2 POINT (-79.79135 35.75225)    1968
#> 3   3  POINT (-78.89238 35.3537)    2030
#> 4   4 POINT (-81.14601 35.71722)    1986
#> 5   5 POINT (-79.42877 36.26948)    1838
#> 6   6  POINT (-79.79379 35.9432)    1903
#> 7   7 POINT (-79.13715 34.53925)    2150
#> 8   8 POINT (-78.22395 33.97557)    2241
#> 9   9 POINT (-76.51494 36.31645)    1881
#> 10 10 POINT (-78.12528 34.16746)    2241
}\if{html}{\out{</div>}}
}

