% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_functions.R
\name{st_extract_within}
\alias{st_extract_within}
\title{Extract information from geometry that a point x falls in}
\usage{
st_extract_within(x, y, var)
}
\arguments{
\item{x}{\code{sf} object consisting of points}

\item{y}{\code{sf} object consisting of geometries}

\item{var}{character of variable name you wish to extract from \code{y}}
}
\description{
For each point in x, identify which geometry in y the point falls within
and extract value of variable \code{var}. This is good, for example, if you have
points and you want to extract information about the county/census tract/etc.
}
\section{Examples}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(sf)
nc <- st_read(system.file("shape/nc.shp", package="sf"))
#> Reading layer `nc' from data source 
#>   `/Library/Frameworks/R.framework/Versions/4.2-arm64/Resources/library/sf/shape/nc.shp' 
#>   using driver `ESRI Shapefile'
#> Simple feature collection with 100 features and 14 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -84.32385 ymin: 33.88199 xmax: -75.45698 ymax: 36.58965
#> Geodetic CRS:  NAD27
x <- st_as_sf(data.frame(id = 1:10, geometry = st_sample(nc, 10)))
x$CNTY_ID = st_extract_within(x, nc, "CNTY_ID")
x
#> Simple feature collection with 10 features and 2 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -82.16131 ymin: 34.44789 xmax: -77.52405 ymax: 36.45466
#> Geodetic CRS:  NAD27
#>    id                   geometry CNTY_ID
#> 1   1 POINT (-81.27872 36.04426)    1874
#> 2   2 POINT (-82.16131 36.04798)    1927
#> 3   3 POINT (-77.85875 36.45466)    1846
#> 4   4 POINT (-80.38286 36.17088)    1900
#> 5   5 POINT (-81.45639 35.56116)    1986
#> 6   6 POINT (-78.63705 36.09166)    1840
#> 7   7 POINT (-77.63797 34.65845)    2146
#> 8   8  POINT (-79.5743 35.36147)    2040
#> 9   9  POINT (-77.52405 35.5089)    2016
#> 10 10  POINT (-78.0643 34.44789)    2185
}\if{html}{\out{</div>}}
}

