% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_functions.R
\name{st_extract_within}
\alias{st_extract_within}
\title{Extract information from geometry that a point x falls in}
\usage{
st_extract_within(x, y, var)
}
\arguments{
\item{x}{\code{sf} object consisting of points}

\item{y}{\code{sf} object consisting of geometries}

\item{var}{character of variable name you wish to extract from \code{y}}
}
\description{
For each point in x, identify which geometry in y the point falls within
and extract value of variable \code{var}. This is good, for example, if you have
points and you want to extract information about the county/census tract/etc.
}
\section{Examples}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(sf)
set.seed(1)
nc <- st_read(system.file("shape/nc.shp", package="sf"))
#> Reading layer `nc' from data source 
#>   `/Library/Frameworks/R.framework/Versions/4.3-arm64/Resources/library/sf/shape/nc.shp' 
#>   using driver `ESRI Shapefile'
#> Simple feature collection with 100 features and 14 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -84.32385 ymin: 33.88199 xmax: -75.45698 ymax: 36.58965
#> Geodetic CRS:  NAD27
x <- st_as_sf(data.frame(id = 1:10, geometry = st_sample(nc, 10)))
x$CNTY_ID = st_extract_within(x, nc, "CNTY_ID")
x
#> Simple feature collection with 10 features and 2 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -82.75835 ymin: 34.93753 xmax: -76.35794 ymax: 36.30275
#> Geodetic CRS:  NAD27
#>    id                   geometry CNTY_ID
#> 1   1 POINT (-79.24443 35.68907)    1973
#> 2   2 POINT (-76.35794 34.94789)    2156
#> 3   3 POINT (-78.46464 34.93753)    2083
#> 4   4 POINT (-78.74558 36.30275)    1840
#> 5   5  POINT (-82.4975 35.21495)    2047
#> 6   6  POINT (-82.75835 35.5431)    1988
#> 7   7  POINT (-78.23211 35.2469)    2029
#> 8   8 POINT (-77.49777 36.18326)    1846
#> 9   9 POINT (-79.91082 35.73617)    1968
#> 10 10 POINT (-77.96082 36.08983)    1913
}\if{html}{\out{</div>}}
}

