% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_functions.R
\name{st_indicator_for_within}
\alias{st_indicator_for_within}
\title{Returns an indicator vector for whether y contains a point of x using st_within}
\usage{
st_indicator_for_within(x, y)
}
\arguments{
\item{x}{\code{sf} object consisting of points}

\item{y}{\code{sf} object consisting of geometries}
}
\description{
Returns an indicator vector for whether y contains a point of x using st_within
}
\section{Examples}{
\if{html}{\out{<div class="r">}}\preformatted{library(sf)
nc <- st_read(system.file("shape/nc.shp", package="sf"))
}\if{html}{\out{</div>}}\preformatted{#> Reading layer `nc' from data source `/Library/Frameworks/R.framework/Versions/4.1/Resources/library/sf/shape/nc.shp' using driver `ESRI Shapefile'
#> Simple feature collection with 100 features and 14 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -84.32385 ymin: 33.88199 xmax: -75.45698 ymax: 36.58965
#> Geodetic CRS:  NAD27
}\if{html}{\out{<div class="r">}}\preformatted{x <- st_as_sf(data.frame(id = 1:10, geometry = st_sample(nc, 10)))
nc$contains_x = st_indicator_for_within(x, nc)
nc
}\if{html}{\out{</div>}}\preformatted{#> Simple feature collection with 100 features and 15 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -84.32385 ymin: 33.88199 xmax: -75.45698 ymax: 36.58965
#> Geodetic CRS:  NAD27
#> First 10 features:
#>     AREA PERIMETER CNTY_ CNTY_ID        NAME  FIPS FIPSNO CRESS_ID BIR74 SID74 NWBIR74 BIR79 SID79 NWBIR79                       geometry contains_x
#> 1  0.114     1.442  1825    1825        Ashe 37009  37009        5  1091     1      10  1364     0      19 MULTIPOLYGON (((-81.47276 3...          0
#> 2  0.061     1.231  1827    1827   Alleghany 37005  37005        3   487     0      10   542     3      12 MULTIPOLYGON (((-81.23989 3...          0
#> 3  0.143     1.630  1828    1828       Surry 37171  37171       86  3188     5     208  3616     6     260 MULTIPOLYGON (((-80.45634 3...          1
#> 4  0.070     2.968  1831    1831   Currituck 37053  37053       27   508     1     123   830     2     145 MULTIPOLYGON (((-76.00897 3...          0
#> 5  0.153     2.206  1832    1832 Northampton 37131  37131       66  1421     9    1066  1606     3    1197 MULTIPOLYGON (((-77.21767 3...          0
#> 6  0.097     1.670  1833    1833    Hertford 37091  37091       46  1452     7     954  1838     5    1237 MULTIPOLYGON (((-76.74506 3...          0
#> 7  0.062     1.547  1834    1834      Camden 37029  37029       15   286     0     115   350     2     139 MULTIPOLYGON (((-76.00897 3...          1
#> 8  0.091     1.284  1835    1835       Gates 37073  37073       37   420     0     254   594     2     371 MULTIPOLYGON (((-76.56251 3...          0
#> 9  0.118     1.421  1836    1836      Warren 37185  37185       93   968     4     748  1190     2     844 MULTIPOLYGON (((-78.30876 3...          0
#> 10 0.124     1.428  1837    1837      Stokes 37169  37169       85  1612     1     160  2038     5     176 MULTIPOLYGON (((-80.02567 3...          0
}
}

