% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_functions.R
\name{st_nearest_distance_rcpp}
\alias{st_nearest_distance_rcpp}
\title{RcppParallel distance to nearest point}
\usage{
st_nearest_distance_rcpp(x, y, unit = "mi")
}
\arguments{
\item{x.}{Object of class sf/sfc or a matrix of coordinates. If a
matrix of coordinates, use column order from \code{st_coordinates()}.}

\item{y.}{Object of class sf/sfc or a matrix of coordinates. If a
matrix of coordinates, use column order from \code{st_coordinates()}.}

\item{unit.}{Either "mi" or "km". Sets the output distance.}
}
\value{
Matrix of two columns. First column is index corresponding to the
closest row of y. The second column is the distance to that element.
}
\description{
RcppParallel distance to nearest point
}
\section{Examples}{
\if{html}{\out{<div class="r">}}\preformatted{library(sf)
nc <- st_read(system.file("shape/nc.shp", package="sf"))
}\if{html}{\out{</div>}}\preformatted{#> Reading layer `nc' from data source `/Library/Frameworks/R.framework/Versions/4.1/Resources/library/sf/shape/nc.shp' using driver `ESRI Shapefile'
#> Simple feature collection with 100 features and 14 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -84.32385 ymin: 33.88199 xmax: -75.45698 ymax: 36.58965
#> Geodetic CRS:  NAD27
}\if{html}{\out{<div class="r">}}\preformatted{st_nearest_distance_rcpp(nc, nc[4:6,])
}\if{html}{\out{</div>}}\preformatted{#> Warning in st_point_on_surface.sfc(x): st_point_on_surface may not give correct results for longitude/latitude data
}\preformatted{#> Warning in st_point_on_surface.sfc(y): st_point_on_surface may not give correct results for longitude/latitude data
}\preformatted{#> → Distance in miles. To use kilometers use option `unit == 'km'`
}\preformatted{#>        [,1]      [,2]
#>   [1,]    2 595012.33
#>   [2,]    2 542861.42
#>   [3,]    2 479360.44
#>   [4,]    1      0.00
#>   [5,]    2      0.00
#>   [6,]    3      0.00
#>   [7,]    1  36643.98
#>   [8,]    3  46677.51
#>   [9,]    2 108062.71
#>  [10,]    2 413287.17
#>  [11,]    2 283705.61
#>  [12,]    2 346666.81
#>  [13,]    2 186097.81
#>  [14,]    2 231889.60
#>  [15,]    2 149547.30
#>  [16,]    2  49127.17
#>  [17,]    1  54513.41
#>  [18,]    2 549074.90
#>  [19,]    2 625425.97
#>  [20,]    1  75639.30
#>  [21,]    3  66652.65
#>  [22,]    2 659311.51
#>  [23,]    2 476221.07
#>  [24,]    2 149395.76
#>  [25,]    2 418396.66
#>  [26,]    2 351077.58
#>  [27,]    2 301333.12
#>  [28,]    3  68871.06
#>  [29,]    2 263485.23
#>  [30,]    2 223441.94
#>  [31,]    2 116070.45
#>  [32,]    2 690631.70
#>  [33,]    2  77981.21
#>  [34,]    2 610205.86
#>  [35,]    2 718222.09
#>  [36,]    2  91830.89
#>  [37,]    2 206359.98
#>  [38,]    2 773111.53
#>  [39,]    2 520586.61
#>  [40,]    2 463932.29
#>  [41,]    2 557065.10
#>  [42,]    2 430954.46
#>  [43,]    2 633559.07
#>  [44,]    3 120435.66
#>  [45,]    1 105122.84
#>  [46,]    2 685491.13
#>  [47,]    2 372690.86
#>  [48,]    2 297688.69
#>  [49,]    2 140889.46
#>  [50,]    2 474901.38
#>  [51,]    2 140508.96
#>  [52,]    2 569630.90
#>  [53,]    2 760839.21
#>  [54,]    2 204305.46
#>  [55,]    2 824499.33
#>  [56,]    1  96807.81
#>  [57,]    3 153032.79
#>  [58,]    2 914336.97
#>  [59,]    2 159574.16
#>  [60,]    2 308892.07
#>  [61,]    2 681611.48
#>  [62,]    2 194775.55
#>  [63,]    2 277133.25
#>  [64,]    2 632485.73
#>  [65,]    2 579814.54
#>  [66,]    2 857454.37
#>  [67,]    2 359746.28
#>  [68,]    2 539289.58
#>  [69,]    2 495121.64
#>  [70,]    2 410154.40
#>  [71,]    2 455572.91
#>  [72,]    2 763166.79
#>  [73,]    2 953169.19
#>  [74,]    2 211288.45
#>  [75,]    2 811433.51
#>  [76,]    2 584702.15
#>  [77,]    2 723047.32
#>  [78,]    2 912696.84
#>  [79,]    2 293132.37
#>  [80,]    3 226866.43
#>  [81,]    2 996065.64
#>  [82,]    2 319677.39
#>  [83,]    2 240043.96
#>  [84,]    2 522155.91
#>  [85,]    2 465437.40
#>  [86,]    2 363735.86
#>  [87,]    1 145089.89
#>  [88,]    2 264329.36
#>  [89,]    2 420322.11
#>  [90,]    2 955692.57
#>  [91,]    2 201122.69
#>  [92,]    2 409460.42
#>  [93,]    2 288480.01
#>  [94,]    2 402726.05
#>  [95,]    1 271978.82
#>  [96,]    2 356786.02
#>  [97,]    2 338604.87
#>  [98,]    2 426504.97
#>  [99,]    2 390232.19
#> [100,]    2 424413.14
}
}

