% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_theme.R
\name{theme_kyle}
\alias{theme_kyle}
\title{Custom ggplot2 theme}
\usage{
theme_kyle(base_size = 14, ...)
}
\arguments{
\item{base_size}{The base_size of the font. Title, axis text, etc. all adjust to base_size. Default is 14.}

\item{...}{Additoinal options passed to \code{ggplot2::theme}}
}
\description{
Custom ggplot2 theme
}
\examples{

ggplot2::ggplot(mtcars) +
  ggplot2::geom_point(ggplot2::aes(x = mpg, y = hp)) +
  theme_kyle(base_size = 18)

ggplot2::ggplot(mtcars) +
  ggplot2::geom_point(ggplot2::aes(x = mpg, y = hp)) +
  ggplot2::facet_wrap(~cyl) +
  ggplot2::labs(
    title = "mtcars Dataset",
    x = "Miles per Gallon", y = "Horsepower"
  ) +
  theme_kyle(base_size = 18)

}
