% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tikzsave.R
\name{tikzsave}
\alias{tikzsave}
\title{Save a tikzpicture and then compile to pdf}
\usage{
tikzsave(filename, plot = last_plot(), packages = NULL, recompile = TRUE, ...)
}
\arguments{
\item{packages}{Vector of strings. Absolute file paths to \code{.sty} files to
include in preamble. Default is my \code{paper.sty} and \code{math.sty} in
\code{inst/tikzsave}.}

\item{recompile}{Logical. Since \code{tikzDevice::tikz} caches effectively,
sometimes the \code{.pdf} file will be newer than the saved \code{.tex} file.
If \code{FALSE}, this will skip recomipilation in these cases.
Default is \code{TRUE}.}

\item{...}{Passed to \code{ggplot2::ggsave}}
}
\value{
Invisibly returns filename
}
\description{
Uses \code{ggplot2::ggsave} with \code{tikzDevice::tikz} to create the tikz figure
and then compiles using \code{latexmk}.
}
